Add data_utils file, moved get_min_imageshape, load_rgbimages, load_masks functions
